+++ server/api/photo/archives/.gitkeep
@@ -1,0 +1,1 @@
+.gitkeep
