+++ server/config.js
@@ -31,6 +31,7 @@ function appMiddleware (app) {
     resave: false,
     saveUninitialized: true
   }));
+app.use(express.static(__dirname + './api/photo/archives'));
   app.use(express.static(__dirname + '/../client'));
 }
