--- server/photo-upload/archives/.gitkeep
@@ -1,1 +1,0 @@
-.gitkeep
