--- server/photo-upload/photoUploadModel.js
@@ -1,1 +1,0 @@
-var photoUploadModel = require('./photoUploadModel.js')
