--- server/photo-upload/photoUploadRoutes.js
@@ -1,6 +1,0 @@
-var photoUploadController = require('./photoUploadController.js');
-
-module.exports = function(router){
-  router.post('/', photoUploadController.postPhoto);
-  router.get('/', photoUploadController.getPhotos);
-};
